+++ application/src/test/java/com/sanctionco/thunder/authentication/basic/BasicAuthConfigurationTest.java
@@ -1,6 +1,7 @@
 package com.sanctionco.thunder.authentication.basic;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.common.collect.Lists;
 import com.google.common.io.Resources;
 import com.sanctionco.thunder.authentication.AuthConfiguration;
 
@@ -33,6 +34,35 @@ class BasicAuthConfigurationTest {
   private static final Validator VALIDATOR = Validators.newValidator();
   private static final YamlConfigurationFactory<AuthConfiguration> FACTORY =
       new YamlConfigurationFactory<>(AuthConfiguration.class, VALIDATOR, OBJECT_MAPPER, "dw");
+
+  @Test
+  void testFromYamlNoKeys() throws Exception {
+    AuthConfiguration configuration = FACTORY.build(new File(Resources.getResource(
+        "fixtures/configuration/auth/basic-auth-no-keys.yaml").toURI()));
+
+    assertTrue(configuration instanceof BasicAuthConfiguration);
+
+    var basicAuthConfiguration = (BasicAuthConfiguration) configuration;
+
+    assertEquals(0, basicAuthConfiguration.getKeys().size());
+  }
+
+  @Test
+  void testFromYamlWithKeys() throws Exception {
+    AuthConfiguration configuration = FACTORY.build(new File(Resources.getResource(
+        "fixtures/configuration/auth/basic-auth-with-keys.yaml").toURI()));
+
+    assertTrue(configuration instanceof BasicAuthConfiguration);
+
+    var basicAuthConfiguration = (BasicAuthConfiguration) configuration;
+
+    assertEquals(2, basicAuthConfiguration.getKeys().size());
+    assertEquals(
+        Lists.newArrayList(
+            new Key("app1", "secret1"),
+            new Key("app2", "secret2")),
+        basicAuthConfiguration.getKeys());
+  }
 
   @Test
   void testRegisterAuthentication() throws Exception {